+++ packages/file_selector/file_selector/test/file_selector_test.dart
@@ -144,80 +144,7 @@ void main() {
     });
   });
 
-  group('getSaveLocation', () {
-    const String expectedSavePath = '/example/path';
-
-    test('works', () async {
-      const int expectedActiveFilter = 1;
-      fakePlatformImplementation
-        ..setExpectations(
-            initialDirectory: initialDirectory,
-            confirmButtonText: confirmButtonText,
-            acceptedTypeGroups: acceptedTypeGroups,
-            suggestedName: suggestedName)
-        ..setPathsResponse(<String>[expectedSavePath],
-            activeFilter: expectedActiveFilter);
-
-      final FileSaveLocation? location = await getSaveLocation(
-        initialDirectory: initialDirectory,
-        confirmButtonText: confirmButtonText,
-        acceptedTypeGroups: acceptedTypeGroups,
-        suggestedName: suggestedName,
-      );
-
-      expect(location?.path, expectedSavePath);
-      expect(location?.activeFilter, acceptedTypeGroups[expectedActiveFilter]);
-    });
-
-    test('works with no arguments', () async {
-      fakePlatformImplementation.setPathsResponse(<String>[expectedSavePath]);
-
-      final FileSaveLocation? location = await getSaveLocation();
-      expect(location?.path, expectedSavePath);
-    });
-
-    test('sets the initial directory', () async {
-      fakePlatformImplementation
-        ..setExpectations(initialDirectory: initialDirectory)
-        ..setPathsResponse(<String>[expectedSavePath]);
-
-      final FileSaveLocation? location =
-          await getSaveLocation(initialDirectory: initialDirectory);
-      expect(location?.path, expectedSavePath);
-    });
-
-    test('sets the button confirmation label', () async {
-      fakePlatformImplementation
-        ..setExpectations(confirmButtonText: confirmButtonText)
-        ..setPathsResponse(<String>[expectedSavePath]);
-
-      final FileSaveLocation? location =
-          await getSaveLocation(confirmButtonText: confirmButtonText);
-      expect(location?.path, expectedSavePath);
-    });
-
-    test('sets the accepted type groups', () async {
-      fakePlatformImplementation
-        ..setExpectations(acceptedTypeGroups: acceptedTypeGroups)
-        ..setPathsResponse(<String>[expectedSavePath]);
-
-      final FileSaveLocation? location =
-          await getSaveLocation(acceptedTypeGroups: acceptedTypeGroups);
-      expect(location?.path, expectedSavePath);
-    });
-
-    test('sets the suggested name', () async {
-      fakePlatformImplementation
-        ..setExpectations(suggestedName: suggestedName)
-        ..setPathsResponse(<String>[expectedSavePath]);
-
-      final FileSaveLocation? location =
-          await getSaveLocation(suggestedName: suggestedName);
-      expect(location?.path, expectedSavePath);
-    });
-  });
-
-  group('getSavePath (deprecated)', () {
+  group('getSavePath', () {
     const String expectedSavePath = '/example/path';
 
     test('works', () async {
@@ -394,7 +321,6 @@ class FakeFileSelector extends Fake
   // Return values.
   List<XFile>? files;
   List<String>? paths;
-  int? activeFilter;
 
   void setExpectations({
     List<XTypeGroup> acceptedTypeGroups = const <XTypeGroup>[],
@@ -413,9 +339,9 @@ class FakeFileSelector extends Fake
     this.files = files;
   }
 
-  void setPathsResponse(List<String> paths, {int? activeFilter}) {
+  // ignore: use_setters_to_change_properties
+  void setPathsResponse(List<String> paths) {
     this.paths = paths;
-    this.activeFilter = activeFilter;
   }
 
   @override
@@ -449,34 +375,11 @@ class FakeFileSelector extends Fake
     String? suggestedName,
     String? confirmButtonText,
   }) async {
-    final FileSaveLocation? result = await getSaveLocation(
-      acceptedTypeGroups: acceptedTypeGroups,
-      options: SaveDialogOptions(
-        initialDirectory: initialDirectory,
-        suggestedName: suggestedName,
-        confirmButtonText: confirmButtonText,
-      ),
-    );
-    return result?.path;
-  }
-
-  @override
-  Future<FileSaveLocation?> getSaveLocation({
-    List<XTypeGroup>? acceptedTypeGroups,
-    SaveDialogOptions options = const SaveDialogOptions(),
-  }) async {
     expect(acceptedTypeGroups, this.acceptedTypeGroups);
-    expect(options.initialDirectory, initialDirectory);
-    expect(options.suggestedName, suggestedName);
-    expect(options.confirmButtonText, confirmButtonText);
-    final String? path = paths?[0];
-    final int? activeFilterIndex = activeFilter;
-    return path == null
-        ? null
-        : FileSaveLocation(path,
-            activeFilter: activeFilterIndex == null
-                ? null
-                : acceptedTypeGroups?[activeFilterIndex]);
+    expect(initialDirectory, this.initialDirectory);
+    expect(suggestedName, this.suggestedName);
+    expect(confirmButtonText, this.confirmButtonText);
+    return paths?[0];
   }
 
   @override