+++ packages/webview_flutter/webview_flutter_wkwebview/test/src/ui_kit/ui_kit_test.mocks.dart
@@ -69,6 +69,21 @@ class MockTestWKWebViewConfigurationHostApi extends _i1.Mock
         returnValueForMissingStub: null,
       );
   @override
+  void setLimitsNavigationsToAppBoundDomains(
+    int? identifier,
+    bool? limit,
+  ) =>
+      super.noSuchMethod(
+        Invocation.method(
+          #setLimitsNavigationsToAppBoundDomains,
+          [
+            identifier,
+            limit,
+          ],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
   void setMediaTypesRequiringUserActionForPlayback(
     int? identifier,
     List<_i3.WKAudiovisualMediaTypeEnumData?>? types,