+++ packages/file_selector/file_selector/example/lib/readme_standalone_excerpts.dart
@@ -39,8 +39,9 @@ class _MyAppState extends State<MyApp> {
   Future<void> saveFile() async {
     // #docregion Save
     const String fileName = 'suggested_name.txt';
-    final String? path = await getSavePath(suggestedName: fileName);
-    if (path == null) {
+    final FileSaveLocation? result =
+        await getSaveLocation(suggestedName: fileName);
+    if (result == null) {
       // Operation was canceled by the user.
       return;
     }
@@ -49,7 +50,7 @@ class _MyAppState extends State<MyApp> {
     const String mimeType = 'text/plain';
     final XFile textFile =
         XFile.fromData(fileData, mimeType: mimeType, name: fileName);
-    await textFile.saveTo(path);
+    await textFile.saveTo(result.path);
     // #enddocregion Save
   }